requirements.txt
discord~=2.5.2
python-dotenv~=1.1.0
Flask~=3.1.1